requirements.txt
streamlit
praw
pandas
numpy
plotly
textblob
wordcloud
transformers
yfinance
networkx
python-dotenv
pytest